pyproject.toml
[project]
name = "leadgen-enricher"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "ddgs>=9.14.0",
    "dnspython>=2.8.0",
    "googlesearch-python>=1.3.0",
    "httpx>=0.28.1",
    "lxml>=6.0.4",
    "playwright>=1.58.0",
    "playwright-stealth>=2.0.3",
    "pydantic>=2.13.1",
    "python-dotenv>=1.2.2",
    "python-whois>=0.9.6",
    "rapidfuzz>=3.14.5",
    "rich>=15.0.0",
    "thefuzz>=0.22.1",
    "whois>=1.20240129.2",
]
